Moved Services.Result to Services.Models
