+++ src/Dfc.CourseDirectory.Services/ProviderService/IProviderService.cs
@@ -1,5 +1,6 @@
 ﻿using System.Collections.Generic;
 using System.Threading.Tasks;
+using Dfc.CourseDirectory.Services.Models;
 using Dfc.CourseDirectory.Services.Models.Providers;
 
 namespace Dfc.CourseDirectory.Services.ProviderService
