+++ src/Dfc.CourseDirectory.Services/ProviderService/ProviderService.cs
@@ -3,6 +3,7 @@ using System.Collections.Generic;
 using System.Net.Http;
 using System.Text;
 using System.Threading.Tasks;
+using Dfc.CourseDirectory.Services.Models;
 using Dfc.CourseDirectory.Services.Models.Providers;
 using Microsoft.Extensions.Logging;
 using Microsoft.Extensions.Options;
